+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpVerifier.kt
@@ -265,9 +265,10 @@ open class OpenId4VpVerifier(
         }
     }
 
-    private fun String.toCreatedRequest(): CreatedRequest = CreatedRequest(this)
-    private fun String.toCreatedRequest(loadRequestObject: suspend (RequestObjectParameters?) -> KmmResult<String>): CreatedRequest =
-        CreatedRequest(this, loadRequestObject)
+    private fun String.toCreatedRequest() = CreatedRequest(this)
+    private fun String.toCreatedRequest(
+        loadRequestObject: suspend (RequestObjectParameters?) -> KmmResult<String>,
+    ) = CreatedRequest(this, loadRequestObject)
 
 
     /**
@@ -346,7 +347,8 @@ open class OpenId4VpVerifier(
         dcqlQuery = if (isDcql) toDCQLQuery() else null,
         presentationDefinition = if (isPresentationExchange)
             toPresentationDefinition(containerJwt, containerSdJwt, rqesFlow) else null,
-        transactionData = if (rqesFlow != PresentationRequestParameters.Flow.UC5) transactionData?.map { it.toBase64UrlJsonString() } else null
+        transactionData = if (rqesFlow != PresentationRequestParameters.Flow.UC5)
+            transactionData?.map { it.toBase64UrlJsonString() } else null
     )
 
     /**
@@ -607,23 +609,10 @@ open class OpenId4VpVerifier(
             val mdocGeneratedNonce = apuDirect?.decodeToString()
                 ?: apuNested?.decodeToString()
                 ?: ""
-            val result = verifier.verifyPresentationIsoMdoc(
+            verifier.verifyPresentationIsoMdoc(
                 input = deviceResponse,
                 verifyDocument = verifyDocument(mdocGeneratedNonce, clientId, responseUrl, expectedNonce)
             )
-            if (result is VerifyPresentationResult.ValidationError) {
-                // This is obviously wrong, but it's implemented that way in EUDIW Ref Wallet for Android
-                // see https://github.com/eu-digital-identity-wallet/eudi-lib-android-wallet-core/pull/153
-                val mdocGeneratedNonce = apuDirect?.encodeToString(Base64UrlStrict)
-                    ?: apuNested?.encodeToString(Base64UrlStrict)
-                    ?: ""
-                verifier.verifyPresentationIsoMdoc(
-                    input = deviceResponse,
-                    verifyDocument = verifyDocument(mdocGeneratedNonce, clientId, responseUrl, expectedNonce)
-                )
-            } else {
-                result
-            }
         }
 
         else -> throw IllegalArgumentException("descriptor.format: $claimFormat")
@@ -640,7 +629,10 @@ open class OpenId4VpVerifier(
         responseUrl: String?,
         expectedNonce: String,
     ): suspend (MobileSecurityObject, Document) -> Boolean = { mso, document ->
-        Napier.d("verifyDocument: mdocGeneratedNonce='$mdocGeneratedNonce', clientId='$clientId', responseUrl='$responseUrl', expectedNonce='$expectedNonce'")
+        Napier.d(
+            "verifyDocument: mdocGeneratedNonce='$mdocGeneratedNonce', clientId='$clientId'," +
+                    " responseUrl='$responseUrl', expectedNonce='$expectedNonce'"
+        )
         val deviceSignature = document.deviceSigned.deviceAuth.deviceSignature ?: run {
             Napier.w("DeviceSignature is null: ${document.deviceSigned.deviceAuth}")
             throw IllegalArgumentException("deviceSignature")